+++ src/main/java/Tuesday/command/MarkCommand.java
@@ -1,19 +1,29 @@
+package Tuesday.command;
+
+import Tuesday.util.Storage;
+import Tuesday.task.Task;
+import Tuesday.util.Ui;
+
 import java.io.*;
 
-public class DeleteCommand extends Command{
-    private int index;
-    public DeleteCommand(String command, int index) {
-        super(command);
-        this.index = index;
+public class MarkCommand extends Command{
+    private String commandType;
+    private int taskIndex;
+    private boolean isMarked;
+    public MarkCommand(String commandType, int taskIndex, boolean isMarked) {
+        super(commandType);
+        this.commandType = commandType;
+        this.taskIndex = taskIndex;
+        this.isMarked = isMarked;
     }
     @Override
     public void execute(Task task, Ui ui, Storage storage) {
-        ui.showDeleteMessage(index - 1);
-        Task.deleteTask(index - 1);
-        this.deleteDataFromFile();
+        Task.taskArrayList.get(this.taskIndex - 1).changeDone(this.isMarked);
+        ui.showMarkMessage(this.taskIndex - 1, this.isMarked);
+        changeDataFromFile();
     }
 
-    public void deleteDataFromFile() {
+    public void changeDataFromFile() {
         int i = 0;
         try {
             BufferedReader br = new BufferedReader(new FileReader("src/main/data/tuesday.txt"));
@@ -22,8 +32,13 @@ public class DeleteCommand extends Command{
                 String line = br.readLine();
 
                 while (line != null) {
-                    if (i != (this.index-1)) {
+                    if (i != (this.taskIndex-1)) {
                         sb.append(line);
+                        sb.append(System.lineSeparator());
+                    } else {
+                        String newData = line.substring(0,4) + Task.taskArrayList.get(this.taskIndex - 1).getDone1()
+                                + line.substring(5);
+                        sb.append(newData);
                         sb.append(System.lineSeparator());
                     }
                     line = br.readLine();
@@ -45,6 +60,7 @@ public class DeleteCommand extends Command{
             System.out.println("Error: IOException");
         }
     }
+
     @Override
     public boolean isExit() {
         return false;
