+++ src/main/java/Tuesday/task/Event.java
@@ -1,3 +1,5 @@
+package Tuesday.task;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
