+++ src/main/java/Tuesday/util/Storage.java
@@ -1,3 +1,10 @@
+package Tuesday.util;
+
+import Tuesday.task.Deadline;
+import Tuesday.task.Event;
+import Tuesday.task.Task;
+import Tuesday.task.ToDo;
+
 import java.io.IOException;
 import java.util.Scanner; // Import the Scanner class
 import java.io.File;
