+++ src/main/java/Tuesday/util/TuesdayException.java
@@ -1,3 +1,5 @@
+package Tuesday.util;
+
 public class TuesdayException extends Exception{
     private String error_message;
     public TuesdayException(String error_message) {
