+++ src/main/java/Tuesday/util/Ui.java
@@ -1,3 +1,7 @@
+package Tuesday.util;
+
+import Tuesday.task.Task;
+
 import java.util.Scanner;
 
 public class Ui {
@@ -19,7 +23,7 @@ public class Ui {
      */
     public void showWelcome() {
         this.showLine();
-        System.out.println("    Hello! I'm Tuesday, a randomly created bot.\n    What can I do for you?");
+        System.out.println("    Hello! I'm Tuesday.Tuesday, a randomly created bot.\n    What can I do for you?");
         this.showLine();
     }
