+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 _______________________________
-Hello! I'm Tuesday, a randomly created bot.
+Hello! I'm Tuesday.Tuesday, a randomly created bot.
 What can I do for you?
 _______________________________
 _______________________________
